+++ CHANGELOG.md
@@ -2,12 +2,18 @@
 
 ## Current Version
 
+### [1.2.0] - 2021-12-05
+
+- `Fix` — Error when only downloading assets
+- `Fix` — When offline checks if there are any assets to download else return OnComplete event
+- `Added` — Option to download binary file formats
+
+## Previous Versions
+
 ### [1.1.1] - 2021-12-05
 
 - `Fix` — Fixed DownloadAssetData function to properly accept Dictionaries
 - `Added` — Added code in sample scene to show how to use dictionaries to download assets
-
-## Previous Versions
 
 ### [1.1.0] - 2021-11-29